speed up container start

+++ docker/opt/start.sh
@@ -126,21 +126,25 @@ zimbra_require_interprocess_security="1"
 zimbra_server_hostname="$HOSTNAME.$DOMAIN"
 INSTALL_PACKAGES="zimbra-core zimbra-ldap zimbra-logger zimbra-mta zimbra-snmp zimbra-store zimbra-apache zimbra-spell zimbra-memcached zimbra-proxy"
 EOF
-##Install the Zimbra Collaboration ##
-echo "Downloading Zimbra Collaboration 8.8.15"
-wget -O /opt/zimbra-install/zimbra-zcs-8.8.15.tar.gz https://files.zimbra.com/downloads/8.8.15_GA/zcs-8.8.15_GA_3829.UBUNTU16_64.20190718141144.tgz
 
-echo "Extracting files from the archive"
-tar xzvf /opt/zimbra-install/zimbra-zcs-8.8.15.tar.gz -C /opt/zimbra-install/
+if [ ! -f /opt/zimbra-install/zimbra-zcs-8.8.15.tar.gz ];
+then
+  ##Install the Zimbra Collaboration ##
+  echo "Downloading Zimbra Collaboration 8.8.15"
+  wget -O /opt/zimbra-install/zimbra-zcs-8.8.15.tar.gz https://files.zimbra.com/downloads/8.8.15_GA/zcs-8.8.15_GA_3829.UBUNTU16_64.20190718141144.tgz
 
-echo "Update package cache"
-apt update
+  echo "Extracting files from the archive"
+  tar xzvf /opt/zimbra-install/zimbra-zcs-8.8.15.tar.gz -C /opt/zimbra-install/
 
-echo "Installing Zimbra Collaboration just the Software"
-cd /opt/zimbra-install/zcs-* && ./install.sh -s < /opt/zimbra-install/installZimbra-keystrokes
+  echo "Update package cache"
+  apt update
 
-echo "Installing Zimbra Collaboration injecting the configuration"
-/opt/zimbra/libexec/zmsetup.pl -c /opt/zimbra-install/installZimbraScript
+  echo "Installing Zimbra Collaboration just the Software"
+  cd /opt/zimbra-install/zcs-* && ./install.sh -s < /opt/zimbra-install/installZimbra-keystrokes
+
+  echo "Installing Zimbra Collaboration injecting the configuration"
+  /opt/zimbra/libexec/zmsetup.pl -c /opt/zimbra-install/installZimbraScript
+fi
 
 su - zimbra -c 'zmcontrol restart'
 echo "You can access now to your Zimbra Collaboration Server"
